Split OPFS tests into separate file (#2964)
## Motivation for the change, related issues

Addresses review feedback from #2961 to split OPFS tests into a separate
file to avoid an artificially large test group.

## Implementation details

- Created `opfs.spec.ts` with all 9 OPFS-related tests (those checking
`browserName !== 'chromium'`)
- Moved `saveSiteViaModal` helper to the new file
- Kept serial execution mode in `opfs.spec.ts` since OPFS storage is
shared at browser level
- Removed serial mode from `website-ui.spec.ts` as remaining tests don't
require it

## Testing Instructions (or ideally a Blueprint)

Run the E2E tests to verify both test files execute correctly:
```bash
npx playwright test packages/playground/website/playwright/e2e/opfs.spec.ts
npx playwright test packages/playground/website/playwright/e2e/website-ui.spec.ts
```
